requirements.txt
astroid==2.4.2
Babel==2.9.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
libcst==0.3.16
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
